Break out dependency of VideoViewManager on ACSCalling Wrap all ACSCalling dependencies through CallingSDKWrapperProtocol Change external CommunicationIdentifier references to CallIdentity class to encapsulate properly
