+++ AzureCommunicationUI/AzureCommunicationUIDemoApp/Sources/RemoteParticipantAvatarHelper.swift
@@ -8,25 +8,25 @@ import AzureCommunicationCommon
 import AzureCommunicationUICalling
 
 struct RemoteParticipantAvatarHelper {
-    private static func getRemoteParticipantId(_ identifier: CommunicationIdentifier) -> String? {
-        switch identifier {
-        case is CommunicationUserIdentifier:
-            return (identifier as? CommunicationUserIdentifier)?.identifier
-        case is UnknownIdentifier:
-            return (identifier as? UnknownIdentifier)?.identifier
-        case is PhoneNumberIdentifier:
-            return (identifier as? PhoneNumberIdentifier)?.phoneNumber
-        case is MicrosoftTeamsUserIdentifier:
-            return (identifier as? MicrosoftTeamsUserIdentifier)?.userId
-        default:
-            return nil
-        }
-    }
+//    private static func getRemoteParticipantId(_ identifier: CommunicationIdentifier) -> String? {
+//        switch identifier {
+//        case is CommunicationUserIdentifier:
+//            return (identifier as? CommunicationUserIdentifier)?.identifier
+//        case is UnknownIdentifier:
+//            return (identifier as? UnknownIdentifier)?.identifier
+//        case is PhoneNumberIdentifier:
+//            return (identifier as? PhoneNumberIdentifier)?.phoneNumber
+//        case is MicrosoftTeamsUserIdentifier:
+//            return (identifier as? MicrosoftTeamsUserIdentifier)?.userId
+//        default:
+//            return nil
+//        }
+//    }
 
-    static func onRemoteParticipantJoined(to callComposite: CallComposite, identifiers: [CommunicationIdentifier]) {
+    static func onRemoteParticipantJoined(to callComposite: CallComposite, identifiers: [CallIdentity]) {
         let avatars = ["cat", "fox", "koala", "monkey", "mouse", "octopus"]
         for identifier in identifiers {
-            let id = getRemoteParticipantId(identifier)
+            let id = identifier.identifier
             let nameIdValue = id != nil ? "\(id?.suffix(4) ?? "")" : ""
             var avatarImage: UIImage?
             var selectedAvatarName = ""
