+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/Manager/AvatarViewManager.swift
@@ -9,7 +9,7 @@ import Combine
 
 protocol AvatarViewManagerProtocol {
     func set(remoteParticipantViewData: ParticipantViewData,
-             for identifier: CommunicationIdentifier,
+             for identifier: CallIdentity,
              completionHandler: ((Result<Void, SetParticipantViewDataError>) -> Void)?)
 }
 
@@ -51,10 +51,10 @@ class AvatarViewManager: AvatarViewManagerProtocol, ObservableObject {
     }
 
     func set(remoteParticipantViewData: ParticipantViewData,
-             for identifier: CommunicationIdentifier,
+             for identifier: CallIdentity,
              completionHandler: ((Result<Void, SetParticipantViewDataError>) -> Void)? = nil) {
         let participantsList = store.state.remoteParticipantsState.participantInfoList
-        guard let idStringValue = identifier.stringValue,
+        guard let idStringValue = identifier.identifier,
               participantsList.contains(where: { $0.userIdentifier == idStringValue })
         else {
             completionHandler?(.failure(SetParticipantViewDataError.participantNotInCall))
