+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Service/Calling/Extension/ACSLocalVideoStreamExtension.swift
@@ -1,0 +1,16 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import AzureCommunicationCalling
+
+extension LocalVideoStream {
+
+    var asVideoStream: VideoStreamInfoModel? {
+        VideoStreamInfoModel(
+            videoStreamIdentifier: source.id,
+            mediaStreamType: mediaStreamType.converted()
+        )
+    }
+}
