+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Service/Calling/Extension/CallingSDKWrapperExtension.swift
@@ -1,0 +1,185 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import AzureCommunicationCalling
+import Foundation
+
+extension CallingSDKWrapper {
+
+    func setupCall() async throws {
+        try await setupCallClientAndDeviceManager()
+    }
+
+    func startCall(isCameraPreferred: Bool, isAudioPreferred: Bool) async throws {
+        logger.debug("Reset Subjects in callingEventsHandler")
+        eventHandler.setupProperties()
+        logger.debug( "Starting call")
+        do {
+            try await setupCallAgent()
+        } catch {
+            throw CallCompositeInternalError.callJoinFailed
+        }
+        try await joinCall(isCameraPreferred: isCameraPreferred, isAudioPreferred: isAudioPreferred)
+    }
+
+    func joinCall(isCameraPreferred: Bool, isAudioPreferred: Bool) async throws {
+        logger.debug( "Joining call")
+        let joinCallOptions = JoinCallOptions()
+
+        if isCameraPreferred,
+           let localVideoStream = localVideoStream {
+            let localVideoStreamArray = [localVideoStream]
+            let videoOptions = VideoOptions(localVideoStreams: localVideoStreamArray)
+            joinCallOptions.videoOptions = videoOptions
+        }
+
+        joinCallOptions.audioOptions = AudioOptions()
+        joinCallOptions.audioOptions?.muted = !isAudioPreferred
+
+        var joinLocator: JoinMeetingLocator
+        if callConfiguration.compositeCallType == .groupCall,
+           let groupId = callConfiguration.groupId {
+            joinLocator = GroupCallLocator(groupId: groupId)
+        } else if let meetingLink = callConfiguration.meetingLink {
+            joinLocator = TeamsMeetingLinkLocator(meetingLink: meetingLink)
+        } else {
+            logger.error("Invalid groupID / meeting link")
+            throw CallCompositeInternalError.callJoinFailed
+        }
+
+        let joinedCall = try await callAgent?.join(with: joinLocator, joinCallOptions: joinCallOptions)
+
+        guard let joinedCall = joinedCall else {
+            logger.error( "Join call failed")
+            throw CallCompositeInternalError.callJoinFailed
+        }
+
+        joinedCall.delegate = callingEventsHandler as? CallingSDKEventsHandler
+        call = joinedCall
+        setupCallRecordingAndTranscriptionFeature()
+    }
+
+    func endCall() async throws {
+        guard call != nil else {
+            throw CallCompositeInternalError.callEndFailed
+        }
+        do {
+            try await call?.hangUp(options: HangUpOptions())
+            logger.debug("Call ended successfully")
+        } catch {
+            logger.error( "It was not possible to hangup the call.")
+            throw error
+        }
+    }
+
+    func muteLocalMic() async throws {
+        guard let call = call else {
+            return
+        }
+
+        do {
+            try await call.mute()
+        } catch {
+            logger.error("ERROR: It was not possible to mute. \(error)")
+            throw error
+        }
+        logger.debug("Mute successful")
+    }
+
+    func unmuteLocalMic() async throws {
+        guard let call = call else {
+            return
+        }
+
+        do {
+            try await call.unmute()
+        } catch {
+            logger.error("ERROR: It was not possible to unmute. \(error)")
+            throw error
+        }
+        logger.debug("Unmute successful")
+    }
+
+    func holdCall() async throws {
+        guard let call = call else {
+            return
+        }
+
+        do {
+            try await call.hold()
+            logger.debug("Hold Call successful")
+        } catch {
+            logger.error("ERROR: It was not possible to hold call. \(error)")
+        }
+    }
+
+    func resumeCall() async throws {
+        guard let call = call else {
+            return
+        }
+
+        do {
+            try await call.resume()
+            logger.debug("Resume Call successful")
+        } catch {
+            logger.error( "ERROR: It was not possible to resume call. \(error)")
+            throw error
+        }
+    }
+
+    private func setupCallRecordingAndTranscriptionFeature() {
+        guard let call = call else {
+            return
+        }
+        let recordingCallFeature = call.feature(Features.recording)
+        let transcriptionCallFeature = call.feature(Features.transcription)
+        eventHandler.assign(recordingCallFeature)
+        eventHandler.assign(transcriptionCallFeature)
+    }
+
+    private func makeCallClient() -> CallClient {
+        let clientOptions = CallClientOptions()
+        let appendingTag = self.callConfiguration.diagnosticConfig.tags
+        let diagnostics = clientOptions.diagnostics ?? CallDiagnosticsOptions()
+        diagnostics.tags.append(contentsOf: appendingTag)
+        clientOptions.diagnostics = diagnostics
+        return CallClient(options: clientOptions)
+    }
+
+    private func setupCallClientAndDeviceManager() async throws {
+        do {
+            let client = makeCallClient()
+            callClient = client
+            let deviceManager = try await client.getDeviceManager()
+            deviceManager.delegate = self
+            self.deviceManager = deviceManager
+        } catch {
+            throw CallCompositeInternalError.deviceManagerFailed(error)
+        }
+    }
+
+    private func setupCallAgent() async throws {
+        guard callAgent == nil else {
+            logger.debug("Reusing call agent")
+            return
+        }
+
+        let options = CallAgentOptions()
+        if let displayName = callConfiguration.displayName {
+            options.displayName = displayName
+        }
+        do {
+            let callAgent = try await callClient?.createCallAgent(
+                userCredential: callConfiguration.credential,
+                options: options
+            )
+            self.logger.debug("Call agent successfully created.")
+            self.callAgent = callAgent
+        } catch {
+            logger.error("It was not possible to create a call agent.")
+            throw error
+        }
+    }
+}
